# swagger.json
{"openapi":"3.0.0","info":{"title":"Todo API","description":"With NextJS by Swagger.\n\n_Please note that some file image that uploaded here will be replaced with static random image, since this API is hosted on Serveless Functions._\n\n`Image Path: https://picsum.photos/`","version":"1.0.2"},"components":{"securitySchemes":{"oauth2":{"type":"apiKey","description":"Standard Authorization header using the Bearer scheme (\"Bearer {token}\")","name":"Authorization","in":"header"}},"schemas":{"LoginDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string","format":"email"},"fullName":{"type":"string"},"accessToken":{"type":"string"}}},"TodoDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"description":{"type":"string"},"deadline":{"type":"string","format":"date-time"},"done":{"type":"boolean"},"snapshootImage":{"type":"string","nullable":true},"userId":{"type":"string","format":"uuid","nullable":true},"user":{"$ref":"#/components/schemas/UserDto"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string","format":"email"},"fullName":{"type":"string"}}},"CreateTodoDto":{"required":["deadline","description"],"type":"object","properties":{"deadline":{"type":"string","format":"date-time","description":"Min value : {datetimeNow}"},"description":{"type":"string","maxLength":100,"description":"Max length : 100"},"snapshootImage":{"type":"string","format":"binary","maxLength":"1024 * 10","description":"Max size : 1024 * 10"}}},"CreateUserDto":{"required":["email","fullName","password"],"type":"object","properties":{"email":{"type":"string","format":"email"},"fullName":{"type":"string"},"password":{"minLenght":6,"type":"string"}}},"EditTodoDto":{"required":["deadline","description","done"],"type":"object","properties":{"deadline":{"type":"string","format":"date-time"},"description":{"type":"string","maxLength":100,"description":"Max length : 100"},"done":{"type":"boolean"},"snapshootImage":{"type":"string","format":"binary","maxSize":"1024 * 10","description":"Max size : 1024 * 10"}}},"SortType":{"type":"string","enum":["asc","desc"]},"LoginUserDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"minLenght":6,"type":"string"}}}}},"paths":{"/api/auth/login":{"post":{"tags":["Auth"],"description":"Login to get to get token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"400":{"description":"Bad Request"}}}},"/api/auth/register":{"post":{"tags":["Auth"],"description":"Register new user to the system","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"Bad Request"}}}},"/api/hello":{"get":{"deprecated":true,"tags":["Hello"],"description":"Returns the hello world","responses":{"200":{"description":"hello world"}}}},"/api/public/todos/{id}":{"get":{"tags":["Todos Public"],"description":"Get spesific todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["Todos Public"],"description":"Edit existing todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/EditTodoDto"}}}},"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request"},"404":{"description":"Not Found"}}},"delete":{"tags":["Todos Public"],"description":"Delete existing todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/api/public/todos":{"get":{"tags":["Todos Public"],"description":"Get list of all todos","parameters":[{"":null,"name":"sortType","in":"query","schema":{"$ref":"#/components/schemas/SortType"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoDto"}}}}},"400":{"description":"Bad Request"}}},"post":{"tags":["Todos Public"],"description":"Create new todo","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"400":{"description":"Bad Request"}}}},"/api/todos/{id}":{"get":{"tags":["Todos"],"description":"Get spesific todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"security":[{"oauth2":[]}]},"put":{"tags":["Todos"],"description":"Edit existing todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/EditTodoDto"}}}},"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"security":[{"oauth2":[]}]},"delete":{"tags":["Todos"],"description":"Delete existing todo","parameters":[{"":null,"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"}},"security":[{"oauth2":[]}]}},"/api/todos":{"get":{"tags":["Todos"],"description":"Get list of all todos","parameters":[{"":null,"name":"sortType","in":"query","schema":{"$ref":"#/components/schemas/SortType"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoDto"}}}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"security":[{"oauth2":[]}]},"post":{"tags":["Todos"],"description":"Create new todo","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoDto"}}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"security":[{"oauth2":[]}]}},"/api/users":{"get":{"deprecated":false,"tags":["Users"],"description":"Show registered user","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}}},"tags":[]}